refactor: rename scheduler.rs to task_scheduler.rs for cron scheduling feature
- Renamed scheduler.rs to task_scheduler.rs for clear distinction
- Updated module declarations and imports in lib.rs
- Updated imports in test files
- Added croner and chrono-tz dependencies for cron scheduling
- Created database migrations for cron_schedules table (PostgreSQL and SQLite)
